fix: reduce linting issues 17370→14686 across all modules
- Remove UTC=UTC self-assignments (PLW0127) — 333 files
- Fix logging f-strings (G004) → % formatting — 850+ instances
- Replace root logger calls (LOG015) with module logger
- Convert str+Enum classes (UP042) to StrEnum
- Remove unused Optional/timezone imports (F401)
- Fix implicit Optional annotations (RUF013) → explicit T | None
- Replace deprecated typing imports (UP035) Dict/List/Set → builtins
- Fix mutable class defaults (RUF012) → ClassVar
- Convert logger.error(exc_info=True) (G201) → logger.exception
- Remove redundant exc arg from logger.exception (TRY401)
- Replace os.path.join/dirname (PTH118/120) → pathlib
- Replace dict()/list() calls (C408) → literals
- Fix asyncio.create_task without reference (RUF006)
- Apply ruff auto-fixes: import sorting, UP017, FURB110/162, C401/408

+++ research/pipeline/online_learning.py
@@ -326,8 +326,8 @@ class AdaptiveBlendWeights:
         primary_prob : Primary model probability.
         online_prob  : Online learner probability.
         """
-        primary_correct = float(int(round(primary_prob)) == label)
-        online_correct = float(int(round(online_prob)) == label)
+        primary_correct = float(round(primary_prob) == label)
+        online_correct = float(round(online_prob) == label)
 
         self._primary_correct.append(primary_correct)
         self._online_correct.append(online_correct)
